@@ -8,70 +8,70 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "ng-high-memory-serve-us-insec": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng serve --proxy-config proxy.conf.prod.us.json --source-map --aot --host 0.0.0.0 --disable-host-check",
-    "ng-high-memory-serve-us": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng serve --proxy-config proxy.conf.prod.us.json --ssl true --source-map --aot --host 0.0.0.0 --disable-host-check",
+    "ng-high-memory-serve-us-insec": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng serve --proxy-config proxy.conf.prod.us.js --source-map --aot --host 0.0.0.0 --disable-host-check",
+    "ng-high-memory-serve-us": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng serve --proxy-config proxy.conf.prod.us.js --ssl true --source-map --aot --host 0.0.0.0 --disable-host-check",
     "ng-high-memory-build-prod": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --prod --base-href /webresource/apps/IoTCloudStarter/ --deploy-url /webresource/apps/IoTCloudStarter/ && cd ./dist/IoTCloudStarter/ && zip -r ./app.zip ./* && cd ../..",
-    "serve_eu": "ng serve --proxy-config proxy.conf.prod.eu.json --ssl true --source-map --aot",
-    "serve_us": "ng serve --proxy-config proxy.conf.prod.us.json --ssl true --source-map --aot",
-    "serve_au": "ng serve --proxy-config proxy.conf.prod.au.json --ssl true --source-map --aot",
-    "serve_eu_cont": "ng serve --proxy-config proxy.conf.prod.eu.json --ssl true --source-map --aot --host 0.0.0.0",
-    "serve_us_cont": "ng serve --proxy-config proxy.conf.prod.us.json --ssl true --source-map --aot --host 0.0.0.0",
-    "serve_au_cont": "ng serve --proxy-config proxy.conf.prod.au.json --ssl true --source-map --aot --host 0.0.0.0",
+    "serve_eu": "ng serve --proxy-config proxy.conf.prod.eu.js --ssl true --source-map --aot",
+    "serve_us": "ng serve --proxy-config proxy.conf.prod.us.js --ssl true --source-map --aot",
+    "serve_au": "ng serve --proxy-config proxy.conf.prod.au.js --ssl true --source-map --aot",
+    "serve_eu_cont": "ng serve --proxy-config proxy.conf.prod.eu.js --ssl true --source-map --aot --host 0.0.0.0",
+    "serve_us_cont": "ng serve --proxy-config proxy.conf.prod.us.js --ssl true --source-map --aot --host 0.0.0.0",
+    "serve_au_cont": "ng serve --proxy-config proxy.conf.prod.au.js --ssl true --source-map --aot --host 0.0.0.0",
     "build_prod_hash": "ng build --prod --base-href /webresource/apps/IoTCloudStarter/index.html --deploy-url /webresource/apps/IoTCloudStarter/ && cd ./dist/IoTCloudStarter/ && zip -r ./app.zip ./* && cd ../..",
     "build_prod": "ng build --prod --base-href /webresource/apps/IoTCloudStarter/ --deploy-url /webresource/apps/IoTCloudStarter/ && cd ./dist/IoTCloudStarter/ && zip -r ./app.zip ./* && cd ../.."
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@tibco-tcstk/tc-core-lib": "^1.2.1",
-    "@tibco-tcstk/tc-forms-lib": "^1.2.1",
-    "@tibco-tcstk/tc-liveapps-lib": "^1.2.1",
-    "@ajsf/bootstrap4": "^0.1.3",
-    "@ajsf/material": "^0.1.3",
-    "@tibco/spotfire-wrapper": "^0.8.0",
-    "angular6-json-schema-form": "^7.3.0",
+    "@ajsf/bootstrap4": "^0.1.4",
+    "@ajsf/material": "^0.1.4",
+    "@angular/animations": "~10.1.6",
+    "@angular/cdk": "~10.2.5",
+    "@angular/common": "~10.1.6",
+    "@angular/compiler": "~10.1.6",
+    "@angular/core": "~10.1.6",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "~10.1.6",
+    "@angular/material": "^10.2.5",
+    "@angular/platform-browser": "~10.1.6",
+    "@angular/platform-browser-dynamic": "~10.1.6",
+    "@angular/router": "~10.1.6",
+    "@tibco-tcstk/tc-core-lib": "^2.0.3-rc.1603889979059",
+    "@tibco-tcstk/tc-forms-lib": "^2.0.3-rc.1603889981195",
+    "@tibco-tcstk/tc-liveapps-lib": "^2.0.3-rc.1603889982561",
+    "@tibco/spotfire-wrapper": "0.13.0",
+    "brace": "^0.11.1",
     "chart.js": "^2.9.3",
     "chartjs-plugin-datalabels": "^0.7.0",
     "chartjs-plugin-doughnutlabel": "^2.0.3",
     "chartjs-plugin-labels": "^1.1.0",
     "chartjs-plugin-streaming": "^1.8.0",
-    "core-js": "^3.6.2",
-    "hammerjs": "^2.0.8",
-    "lodash": "^4.17.19",
-    "lodash-es": "^4.17.15",
-    "ng2-charts": "2.3.0",
-    "ngx-color-picker": "^8.2.0",
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1",
-    "brace": "^0.11.1",
+    "core-js": "^3.6.5",
     "fuzzy": "^0.1.3",
-    "ng2-google-charts": "^4.0.0",
+    "hammerjs": "^2.0.8",
+    "heatmap.js": "git+https://github.com/sanuradhag/heatmap.js.git",
+    "lodash": "^4.17.20",
+    "lodash-es": "^4.17.15",
+    "minimist": ">=1.2.3",
     "mixin-deep": ">=1.3.2",
-    "minimist": ">=1.2.3"
+    "ng2-charts": "2.4.2",
+    "ng2-google-charts": "^4.0.0",
+    "ngx-color-picker": "^10.1.0",
+    "ngx-heatmap": "^1.0.2",
+    "rxjs": "^6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular/cli": "^8.3.26",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@tibco-tcstk/component-template": "^1.2.4",
-    "@types/node": "^12.12.24",
-    "child_process": "^1.0.2",
-    "codelyzer": "^5.2.1",
-    "ts-node": "~8.5.2",
-    "tslint": "~5.20.1",
-    "typescript": "~3.4.5",
+    "@angular-devkit/build-angular": "^0.1001.7",
+    "@angular/cli": "^10.1.7",
+    "@angular/compiler-cli": "^10.1.6",
+    "@angular/language-service": "~10.1.6",
+    "@tibco-tcstk/component-template": "^1.2.5",
+    "@types/node": "^14.11.8",
+    "codelyzer": "^6.0.1",
+    "ts-node": "~8.10.2",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7",
     "del": "^5.1.0",
     "file-system": "^2.2.2",
     "fs-extra": "^8.1.0",
