@@ -13,6 +13,7 @@ export class IotDashboardComponent implements OnInit {
   // Spotfire configuration
   public spotfireServer: string;
   public analysisPath: string;
+  public version: string;
   public allowedPages: string[];
   public activePage: string;
   public markingOn = {};
@@ -48,11 +49,17 @@ export class IotDashboardComponent implements OnInit {
     // https://spotfire-next.cloud.tibco.com/spotfire/wp/analysis?file=/Users/b5zl5zgs2jshn2xyyess4gzqufcuue6q/Public/Product%20Registration%2006
 
 
-    this.spotfireServer = 'https://spotfire-next.cloud.tibco.com';
-    this.analysisPath = '/Users/b5zl5zgs2jshn2xyyess4gzqufcuue6q/Public/Product Registration 06';
+    // this.spotfireServer = 'https://spotfire-next.cloud.tibco.com';
+    // this.analysisPath = '/Users/b5zl5zgs2jshn2xyyess4gzqufcuue6q/Public/Product Registration 06';
 
     this.spotfireServer = 'https://ec2-3-223-106-221.compute-1.amazonaws.com';
     this.analysisPath = '/Anonymous/MSAnomalyDetection';
+    this.version = '10.7'
+
+    // https://demo.spotfire.cloud.tibco.com/spotfire/wp/analysis?file=/Public/Working%20Demos/Wafermap%20Pattern%20Recognition%20-%20Air%20Demo&waid=QHnEu1TdLku_b9yhuVaq3-120416f172FBqm&wavid=0
+
+    // this.spotfireServer = 'https://demo.spotfire.cloud.tibco.com';
+    // this.analysisPath = '/Public/Working Demos/Wafermap Pattern Recognition-Air Demo';
 
     this.activePage = '0';
     this.parameters = "Test";
