@@ -3,8 +3,10 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { EdgeService } from '../../services/edge/edge.service';
 import { GraphService } from '../../services/graph/graph.service';
 import { Device, TSReading, Resource, Gateway, ModelConfig } from '../../shared/models/iot.model';
-import { MatPaginator, MatSort, MatTableDataSource, MatSnackBar } from '@angular/material';
 import { SelectionModel } from '@angular/cdk/collections';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
 
 export interface SelectItem {
   value: string;
@@ -34,7 +36,21 @@ export class IotMlModelsComponent implements OnInit {
   actionResourcesDataSource = new MatTableDataSource<Resource>();
 
   mlmodels: SelectItem[] = [
-    { value: 'AnomalyDetection', viewValue: 'Anomaly Detection' }
+    { value: 'ms|anomaly_detection|AnomalyDetection', viewValue: 'Anomaly Detection' },
+    { value: 'tibco|pattern_recognition|PatternRecognition', viewValue: 'Wafer Pattern Recognition' },
+    { value: 'nvidia|image_recognition|alexnet', viewValue: 'Object Detection - AlexNet' },
+    { value: 'nvidia|image_recognition|googlenet', viewValue: 'Object Detection - GoogleNet' },
+    { value: 'nvidia|image_recognition|googlenet-12', viewValue: 'Object Detection - GoogleNet-12' },
+    { value: 'nvidia|image_recognition|resnet-18', viewValue: 'Object Detection - Resnet-18' },
+    { value: 'nvidia|image_recognition|resnet-50', viewValue: 'Object Detection - Resnet-50' },
+    { value: 'nvidia|image_recognition|resnet-101', viewValue: 'Object Detection - Resnet-101' },
+    { value: 'nvidia|image_recognition|resnet-152', viewValue: 'Object Detection - Resnet-152' },
+    { value: 'nvidia|image_recognition|vgg-16', viewValue: 'Object Detection - VGG-16' },
+    { value: 'nvidia|image_recognition|vgg-19', viewValue: 'Object Detection - VGG-19' },
+    { value: 'nvidia|image_recognition|inception_v4', viewValue: 'Object Detection - Inception-v4' },
+    { value: 'tfserving|image_recognition|rcnnresnet', viewValue: 'Object Detection - RCNN Resnet' },
+    { value: 'tfserving|image_recognition|resnet', viewValue: 'Object Detection - Resnet' },
+    { value: 'tfserving|image_recognition|inception', viewValue: 'Object Detection - Imagenet Inception' }
   ];
 
 
@@ -145,7 +161,7 @@ export class IotMlModelsComponent implements OnInit {
       modified: row.modified,
       uid: row.uid
     }, { emitEvent: true });
-    
+
   }
 
   onDeviceSelected(event) {
@@ -161,16 +177,16 @@ export class IotMlModelsComponent implements OnInit {
   onResourceSelected(event) {
 
   }
- 
+
 
   resetModelForm() {
     console.log("Resetting modelConfig form");
-    
+
     this.modelForm.reset({
     }, { emitEvent: false });
 
     console.log("Form after resetting: ", this.modelForm);
-    
+
   }
 
   addModelConfig() {
@@ -234,13 +250,13 @@ export class IotMlModelsComponent implements OnInit {
     let idx = 0;
 
     for (let i = 0; i < this.devicesDataSource.data.length; i++) {
-       
+
       if (this.devicesDataSource.data[i].name == name) {
         idx = i;
         break;
-      } 
+      }
     }
-    
+
     return idx;
   }
 
