@@ -1,6 +1,5 @@
 import {LoginComponent} from '../routes/login/login.component';
 import {
-  AuthGuard,
   GeneralConfigResolver,
   LoginPrefillResolver,
   TibcoCloudErrorComponent,
@@ -8,8 +7,10 @@ import {
 import {StarterAppComponent} from '../routes/starter-app/starter-app.component';
 import {
   ClaimsResolver,
+  AuthGuard
 } from '@tibco-tcstk/tc-liveapps-lib';
 import {STARTER_APP_ROUTES, STARTER_APP_PROVIDERS, HOME_ROUTE } from './starter-app-route-config/starter-app-route-config';
+import { LoginOauthComponent } from '../routes/login-oauth/login-oauth.component';
 
 export const CORE_ROUTES = [
       {
@@ -18,6 +19,10 @@ export const CORE_ROUTES = [
         resolve: {
           loginPrefill: LoginPrefillResolver
         }
+      },
+      {
+        path: 'login-oauth',
+        component: LoginOauthComponent
       },
       {
         path: 'errorHandler/:errorCode/:errorData',
